Improvement for events/ endpoint
event/ used to load all events in database into memory and then applied
filters and limits in python code, which might lead to a memory problems
if there were more events. This has been fixed by performing filtering
and applying limits on the database side.

Alongside with that events/ now display all upcoming events, and does
not display the ones in the past.